webapp/DataEntry_v2/requirements.txt
amqp==2.6.1
asn1crypto==0.24.0
billiard==3.6.1.0
celery==4.3.0
certifi==2018.8.24
cffi==1.15.1
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.6.1
decorator==4.3.0
Django==1.11.23
django-appconf==1.0.3
django-easy-select2==1.5.6
django-parler==1.9.2
django-select2-forms==2.1.0
django-sortedm2m==2.0.0
docopt==0.6.2
EditorConfig==0.12.1
entrypoints==0.3
enum34==1.1.6
feedparser==5.2.1
gunicorn==19.9.0
idna==2.6
importlib-metadata==0.23
ipaddress==1.0.17
jsbeautifier==1.6.4
keyring==17.1.1
keyrings.alt==3.1.1
kombu==4.6.11
meld3==1.0.2
more-itertools==5.0.0
musicbrainzngs==0.6
mutagen==1.40.0
mysqlclient==1.4.6
ngxtop==0.0.3
numpy==1.16.2
olefile==0.46
pathlib==1.0.1
pathlib2==2.3.5
Pillow==5.4.1
psutil==5.5.1
pyasn1==0.4.2
pycairo==1.18.2
pycparser==2.21
pycrypto==2.6.1
pyinotify==0.9.6
pyOpenSSL==19.0.0
pyparsing==2.4.7
pytesseract==0.3.0
python-dateutil==2.7.3
python-magic==0.4.27
pytz==2019.2
pyxdg==0.25
pyzbar==0.1.8
redis==3.3.11
requests==2.21.0
scandir==1.10.0
SecretStorage==2.3.1
six==1.12.0
tabulate==0.8.9
urllib3==1.24
uTidylib==0.6
vine==1.3.0
zipp==0.6.0
